secret fix

+++ CS50_News/views.py
@@ -10,7 +10,7 @@ from django.core.paginator import Paginator
 from django.template.defaulttags import register
 from django.core import serializers
 from datetime import datetime, timezone
-from allauth.account.decorators import reauthentication_required, login_required, verified_email_required, secure_admin_login
+from allauth.account.decorators import login_required, verified_email_required, secure_admin_login, reauthentication_required, reauthentication
 from django.contrib.admin.views.decorators import staff_member_required
 from django.core.files.uploadedfile import InMemoryUploadedFile
 from django.db.models import Count
@@ -18,6 +18,7 @@ from PIL import Image
 from io import BytesIO
 from sys import getsizeof
 from math import isclose
+from allauth.account.views import ConfirmEmailVerificationCodeView
 from taggit.models import Tag
 import pyotp
 import json
@@ -26,8 +27,7 @@ from .models import User, New
 from .utils import send_otp, Rescore, short_category
 
 
-# Create your views here.
-def index(request, cat=None, sub=None):
+def index(request, cat=None, sub=None, key=None):
     Rescore(New.objects.all())
     parent = None
     sub_categories = None
@@ -58,7 +58,8 @@ def index(request, cat=None, sub=None):
         "only": only,
         "featured": featured,
         "subs": sub_categories,
-        "parent": cat
+        "parent": cat,
+        "key": key
     })
 
 def search(request):
@@ -80,9 +81,9 @@ def logout_view(request):
 
 @secure_admin_login
 def admin_view(request):
-    news = New.objects.filter(auther=request.user)
     return render(request, "CS50_News/admin.html", {
-        "news": news
+        "news": New.objects.filter(auther=request.user),
+        "users": User.objects.all()
     })
 
 @secure_admin_login
@@ -98,53 +99,50 @@ def add_new(request):
         short_name = short_category(sub_category)
         if not sub_category in New.SUB_CATEGORIES[category[0]].values():
             return HttpResponse("the provided sub category isn't part of the category provided", status=400)
-        #try:
-        with Image.open(image) as img:
-            w, h = img.size
-            aspect_ratio = 16 / 9
-            if isclose(w / h, aspect_ratio):
-                newh = h
-                neww = int(h * aspect_ratio)
+        try:
+            with Image.open(image) as img:
+                w, h = img.size
+                aspect_ratio = 16 / 9
+                if isclose(w / h, aspect_ratio):
+                    newh = h
+                    neww = int(h * aspect_ratio)
 
-                if neww > w:
-                    neww = w
-                    newh = int(w / aspect_ratio)
-                
-                left = (w - neww) / 2
-                top = (h - newh) / 2
-                right = (w + neww) / 2
-                bottom = (h + newh) / 2
-                img = img.crop((left, top, right, bottom))
+                    if neww > w:
+                        neww = w
+                        newh = int(w / aspect_ratio)
+                    
+                    left = (w - neww) / 2
+                    top = (h - newh) / 2
+                    right = (w + neww) / 2
+                    bottom = (h + newh) / 2
+                    img = img.crop((left, top, right, bottom))
 
-                img_io = BytesIO()
-                img.save(img_io, format='PNG')  # Change format if necessary
-                img_io.seek(0)
+                    img_io = BytesIO()
+                    img.save(img_io, format='PNG')  # Change format if necessary
+                    img_io.seek(0)
 
-                cropped_image_file = InMemoryUploadedFile(
-                img_io,
-                field_name="image",
-                name=image.name,
-                content_type="image/png",
-                size=getsizeof(img_io),
-                charset=None
-                )
-                new = New(headline=headline, sub_headline=sub_headline, image=cropped_image_file, content=content, auther=request.user, category=category[0], sub_category=short_name)
-                new.save()
-            else:
-                new = New(headline=headline, sub_headline=sub_headline, image=image, content=content, auther=request.user, category=category[0], sub_category=short_name)
-                new.save()
-
-        #except:
-
-         #   return HttpResponse("could not open the image", status=400)
-        #try:
-        tags = tags.split(",")
-        print(tags)
-        for tag in tags:
-            new.tags.add(tag)
-        new.save()
-        #except:
-         #  return HttpResponse({"could not create the article"}, status=400)
+                    cropped_image_file = InMemoryUploadedFile(
+                    img_io,
+                    field_name="image",
+                    name=image.name,
+                    content_type="image/png",
+                    size=getsizeof(img_io),
+                    charset=None
+                    )
+                    new = New(headline=headline, sub_headline=sub_headline, image=cropped_image_file, content=content, auther=request.user, category=category[0], sub_category=short_name)
+                    new.save()
+                else:
+                    new = New(headline=headline, sub_headline=sub_headline, image=image, content=content, auther=request.user, category=category[0], sub_category=short_name)
+                    new.save()
+        except:
+           return HttpResponse("could not open the image", status=400)
+        try:
+            tags = tags.split(",")
+            for tag in tags:
+                new.tags.add(tag)
+            new.save()
+        except:
+           return HttpResponse({"could not create the article"}, status=400)
         return HttpResponseRedirect(reverse("staff"))
     suggestions = Tag.objects.all().annotate(num_tags=Count("new")).order_by("-num_tags")
     return render(request, "CS50_News/editor.html", {
@@ -166,42 +164,51 @@ def otp_view(request):
         otp = request.POST.get("otp")
         otp_secret_key = request.session["otp_secret_key"]
         otp_valid_util = request.session["otp_valid_until"]
-        if otp_secret_key and otp_secret_key:
-            valid_until = datetime.fromisoformat(otp_valid_util)
-            if valid_until > datetime.now():
-                totp = pyotp.TOTP(otp_secret_key, interval=360)
-                if totp.now() == otp:
-                    user = User.objects.get(id=request.user.id)
-                    user.otp_date = datetime.now(timezone.utc)
-                    return HttpResponse(status=200)
-                return HttpResponse(status=401)
-            return HttpResponse(status=408)
-        return HttpResponse(status=500)
+        if not (otp_secret_key and otp_secret_key):
+            return HttpResponse(status=500)
+        valid_until = datetime.fromisoformat(otp_valid_util)
+        if not valid_until > datetime.now():
+                    return HttpResponse(status=408)
+        totp = pyotp.TOTP(otp_secret_key, interval=360)
+        if not totp.now() == otp:
+            return HttpResponse(status=401)
+        user = User.objects.get(id=request.user.id)
+        user.otp_date = datetime.now(timezone.utc)
+        return HttpResponse(status=200)
     send_otp(request)
     return HttpResponse(status=200)
 
-@reauthentication_required
-def Delete(request):
+def delete_account(request):
+    if (
+            request.user.is_anonymous
+            or not reauthentication.did_recently_authenticate(request)
+        ):
+        return HttpResponseRedirect(reverse("index"))
     if request.method == "POST":
         try:
-            user = User.objects.get(id=request.user.id)
+            user = request.user
             user.delete()
-            return redirect("/")
+            return HttpResponseRedirect(reverse("index"))
         except User.DoesNotExist:
             return HttpResponse({"error":"There is no such account, please reload the page and try again"}, status=500)
-    return HttpResponse({"error":"bad request method"}, status=400)
+    return HttpResponse({"error":"bad request method"}, status=405)
 
-@reauthentication_required
 def accountEdit(request):
+    if (
+        request.user.is_anonymous
+        or not reauthentication.did_recently_authenticate(request)
+        ):
+        return HttpResponseRedirect(reverse("index"), status=403)
     if request.method == "POST":
         data = json.loads(request.body)
         user = User.objects.get(id=request.user.id)
-        user.username = data.get("username")
-        user.first_name = data.get("first-name")
-        user.last_name = data.get("last-name")
-        user.save()
+        try:
+            user.username = data.get("username")
+            user.save()
+        except IntegrityError:
+            return HttpResponse(status=400)
         return HttpResponse(status=200)
-    return HttpResponse(status=403)
+    return HttpResponse(status=405)
 
 def crop(request):
     if request.method == 'POST':
@@ -221,3 +228,31 @@ def new(request, cat, id):
         "relatedNews": related,
         "tags": news.tags.names()
     })
+
+def reset(request, key):
+    return render(request, "CS50_News/reset.html", {
+        "key": key
+    })
+
+def reauthenticate_decision(request):
+    if (
+            request.user.is_anonymous
+            or not reauthentication.did_recently_authenticate(request)
+            ):
+        return HttpResponse(status=401)
+    else:
+        return HttpResponse(status=200)
+    
+def check(request):
+    return render(request, "CS50_News/index.html")
+
+def settings(request):
+    return render(request, "CS50_News/setting.html")
+
+class CustomConfirmCode(ConfirmEmailVerificationCodeView):
+    def form_valid(self, form):
+        response = super().form_valid(form)
+        return JsonResponse({
+            "message": "Email confirmed successfully!",
+            "redirect_url": reverse('home') + "?auth=confirm-email"
+        })
